Etc: PR 리뷰 반영 - ViewHolder 파일 분리, 코드 오타 수정 및 불필요한 코드 삭제
