+++ Pyonsnal-Color/Pyonsnal-Color/Favorite/FavoriteViewController.swift
@@ -21,7 +21,7 @@ protocol FavoritePresentableListener: AnyObject {
     var isPagingEnabled: Bool { get }
 }
 
-enum FavoriteTab: Int {
+enum FavoriteTab: Int, CaseIterable {
     case product = 0
     case event
     
@@ -38,7 +38,7 @@ enum FavoriteTab: Int {
 final class FavoriteViewController: UIViewController,
                                     FavoritePresentable,
                                     FavoriteViewControllable {
-
+    // MARK: Interface
     enum Size {
         static let headerViewHeight: CGFloat = 48
         static let stackViewHeight: CGFloat = 40
@@ -53,8 +53,9 @@ final class FavoriteViewController: UIViewController,
     }
     
     weak var listener: FavoritePresentableListener?
+    
+    // MARK: Private Property
     private let viewHolder: ViewHolder = .init()
-
     private var products = [FavoriteTab: [any ProductConvertable]]()
     private var scrollIndex = CurrentValueSubject<Int, Never>(0)
     private var cancellable = Set<AnyCancellable>()
@@ -68,6 +69,7 @@ final class FavoriteViewController: UIViewController,
         fatalError("init(coder:) has not been implemented")
     }
     
+    // MARK: - View Life Cycle
     override func viewWillAppear(_ animated: Bool) {
         listener?.requestFavoriteProducts()
     }
@@ -87,11 +89,13 @@ final class FavoriteViewController: UIViewController,
         listener?.deleteAllProducts()
     }
     
+    // MARK: - Public Method
     func updateProducts(products: [any ProductConvertable], tab: FavoriteTab) {
         self.products[tab] = products
         viewHolder.collectionView.reloadData()
     }
     
+    // MARK: - Private Method
     private func configureTabBarItem() {
         tabBarItem.setTitleTextAttributes([.font: UIFont.label2], for: .normal)
         tabBarItem = UITabBarItem(
@@ -178,136 +182,6 @@ final class FavoriteViewController: UIViewController,
         }.store(in: &cancellable)
         
     }
-    
-    final class ViewHolder: ViewHolderable {
-        let titleNavigationView: TitleNavigationView = {
-            let view = TitleNavigationView()
-            view.updateTitleLabel(with: Text.tabBarItem)
-            return view
-        }()
-        
-        private let stackView: UIStackView = {
-            let stackView = UIStackView()
-            stackView.distribution = .fillEqually
-            stackView.axis = .horizontal
-            return stackView
-        }()
-        
-        private let dividerView: UIView = {
-            let view: UIView = .init(frame: .zero)
-            view.backgroundColor = .gray200
-            return view
-        }()
-        
-        private let productSubView: UIView = {
-            let view = UIView()
-            return view
-        }()
-        
-        let productTabButton: UIButton = {
-            let button = UIButton()
-            button.setText(with: Text.productTab)
-            button.titleLabel?.font = .label1
-            button.setTitleColor(.black, for: .selected)
-            button.setTitleColor(.gray400, for: .normal)
-            return button
-        }()
-        
-        private let eventSubView: UIView = {
-            let view = UIView()
-            return view
-        }()
-        
-        let eventTabButton: UIButton = {
-            let button = UIButton()
-            button.setText(with: Text.eventTab)
-            button.titleLabel?.font = .label1
-            button.setTitleColor(.black, for: .selected)
-            button.setTitleColor(.gray400, for: .normal)
-            return button
-        }()
-        
-        let productUnderBarView: UIView = {
-            let view = UIView()
-            view.backgroundColor = .black
-            return view
-        }()
-        
-        let eventUnderBarView: UIView = {
-            let view = UIView()
-            view.backgroundColor = .black
-            return view
-        }()
-        
-        let collectionView: UICollectionView = {
-            let layout = UICollectionViewFlowLayout()
-            layout.scrollDirection = .horizontal
-            let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
-            collectionView.isPagingEnabled = true
-            collectionView.bounces = false
-            collectionView.showsHorizontalScrollIndicator = false
-            return collectionView
-        }()
-        
-        func place(in view: UIView) {
-            view.addSubview(titleNavigationView)
-            view.addSubview(stackView)
-            view.addSubview(dividerView)
-            stackView.addArrangedSubview(productSubView)
-            stackView.addArrangedSubview(eventSubView)
-            productSubView.addSubview(productTabButton)
-            productSubView.addSubview(productUnderBarView)
-            eventSubView.addSubview(eventTabButton)
-            eventSubView.addSubview(eventUnderBarView)
-            view.addSubview(collectionView)
-        }
-        
-        func configureConstraints(for view: UIView) {
-            titleNavigationView.snp.makeConstraints {
-                $0.top.equalTo(view.safeAreaLayoutGuide.snp.top)
-                $0.height.equalTo(Size.headerViewHeight)
-                $0.leading.trailing.equalToSuperview()
-            }
-            
-            stackView.snp.makeConstraints {
-                $0.leading.equalToSuperview().offset(.spacing16)
-                $0.trailing.equalToSuperview().inset(.spacing16)
-                $0.top.equalTo(titleNavigationView.snp.bottom)
-                $0.height.equalTo(Size.stackViewHeight)
-            }
-            
-            dividerView.snp.makeConstraints { make in
-                make.height.equalTo(Size.dividerViewHeight)
-                make.bottom.equalTo(stackView.snp.bottom)
-                make.leading.trailing.equalToSuperview()
-            }
-            
-            productTabButton.snp.makeConstraints {
-                $0.top.leading.trailing.bottom.equalToSuperview()
-            }
-            
-            eventTabButton.snp.makeConstraints {
-                $0.top.leading.trailing.bottom.equalToSuperview()
-            }
-            
-            productUnderBarView.snp.makeConstraints {
-                $0.leading.trailing.bottom.equalToSuperview()
-                $0.height.equalTo(Size.underBarHeight)
-            }
-            
-            eventUnderBarView.snp.makeConstraints {
-                $0.leading.trailing.bottom.equalToSuperview()
-                $0.height.equalTo(Size.underBarHeight)
-            }
-            
-            collectionView.snp.makeConstraints {
-                $0.top.equalTo(stackView.snp.bottom)
-                $0.leading.trailing.equalToSuperview()
-                $0.bottom.equalTo(view.safeAreaLayoutGuide.snp.bottom)
-            }
-        }
-
-    }
 
 }
 
@@ -334,7 +208,7 @@ extension FavoriteViewController: ProductPresentable {
 // MARK: - UICollectionViewDataSource
 extension FavoriteViewController: UICollectionViewDataSource {
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return 2
+        return FavoriteTab.allCases.count
     }
 
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
